接口调试

+++ src/main/java/com/github/zhgxun/talk/service/impl/BookServiceImpl.java
@@ -21,7 +21,7 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public BookEntity add(BookEntity entity) {
-        if (bookDao.findOne(0, entity.getTitle()) != null) {
+        if (bookDao.findOne(0, entity.getCategoryId(), entity.getTitle()) != null) {
             throw new NormalException("该书已经存在");
         }
         bookDao.add(entity);
@@ -29,23 +29,31 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public BookEntity findOne(int id, String title) {
-        return bookDao.findOne(id, title);
+    public BookEntity findOne(int id, int categoryId, String title) {
+        return bookDao.findOne(id, categoryId, title);
     }
 
     @Override
-    public List<BookEntity> any(String title, String author, String nickName) {
-        return bookDao.any(title, author, nickName);
+    public List<BookEntity> findAny(int categoryId, String title, String author, String nickName) {
+        return bookDao.findAny(categoryId, title, author, nickName);
     }
 
     @Override
     public int update(int id, String url, String description, int playCount) {
-        return bookDao.update(id, url, description, playCount);
+        BookEntity entity = bookDao.findOne(id, 0, null);
+        if (entity == null) {
+            throw new NormalException("图书不存在");
+        }
+        return bookDao.update(entity.getId(), url, description, playCount);
     }
 
     @Override
     public int delete(int id) {
-        itemDao.delete(0, id);
+        BookEntity entity = bookDao.findOne(id, 0, null);
+        if (entity == null) {
+            throw new NormalException("图书不存在");
+        }
+        itemDao.delete(0, entity.getId());
         return bookDao.delete(id);
     }
 }
